assignment/requirements.txt
click>=8.0.0
psycopg2-binary>=2.9.0
requests>=2.28.0
pytest>=7.0.0